requirements.txt
altgraph==0.17.4
asgiref==3.8.1
astroid==3.3.5
attrs==25.3.0
auto-py-to-exe==2.45.1
autopep8==2.3.1
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cx-Freeze==7.2.8
cx-Logging==3.2.1
dill==0.3.9
Django==3.2.2
django-environ==0.12.0
djangorestframework==3.12.4
Eel==0.18.1
et-xmlfile==2.0.0
exceptiongroup==1.2.2
flask==3.1.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==20.0.4
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.5
lief==0.16.2
lxml==5.3.1
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
pep8==1.7.1
pillow==11.0.0
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
psycopg2-binary==2.9.9
py2exe==0.13.0.2
pycodestyle==2.12.1
pycparser==2.22
pygame==2.6.1
pygbag==0.9.2
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pylint==3.3.1
pyparsing==3.2.1
pyshorteners==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.4
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tomli==2.0.2
tomlkit==0.13.2
trio==0.29.0
trio-websocket==0.12.2
typing-extensions==4.12.2
tzdata==2025.1
UNKNOWN==0.0.0
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
werkzeug==3.1.3
wsproto==1.2.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
